Disk cache ongoing

+++ src/main/java/cc/kostic/gec/instrument/Greeks.java
@@ -2,37 +2,39 @@ package cc.kostic.gec.instrument;
 
 import org.json.JSONObject;
 
+import java.math.BigDecimal;
+
 public class Greeks {
 	
-	private final Double	delta;		// "delta":0
-	private final Double	gamma;		// "gamma":0
-	private final Double	vega;		// "vega":0.00001
-	private final Double	theta;		// "theta":-0.00003
-	private final Double	rho;		// "rho":0
+	private final BigDecimal	delta;		// "delta":0
+	private final BigDecimal gamma;		// "gamma":0
+	private final BigDecimal	vega;		// "vega":0.00001
+	private final BigDecimal	theta;		// "theta":-0.00003
+	private final BigDecimal	rho;		// "rho":0
 
 	
 	public Greeks(JSONObject data) {
-		this.delta = data.optBigDecimal("delta", null).doubleValue();
-		this.gamma = data.optBigDecimal("gamma", null).doubleValue();
-		this.vega = data.optBigDecimal("vega", null).doubleValue();
-		this.theta = data.optBigDecimal("theta", null).doubleValue();
-		this.rho = data.optBigDecimal("rho", null).doubleValue();
+		this.delta = data.optBigDecimal("delta", null);
+		this.gamma = data.optBigDecimal("gamma", null);
+		this.vega = data.optBigDecimal("vega", null);
+		this.theta = data.optBigDecimal("theta", null);
+		this.rho = data.optBigDecimal("rho", null);
 	}
 	
 	
-	public Double getDelta() {
+	public BigDecimal getDelta() {
 		return delta;
 	}
-	public Double getGamma() {
+	public BigDecimal getGamma() {
 		return gamma;
 	}
-	public Double getVega() {
+	public BigDecimal getVega() {
 		return vega;
 	}
-	public Double getTheta() {
+	public BigDecimal getTheta() {
 		return theta;
 	}
-	public Double getRho() {
+	public BigDecimal getRho() {
 		return rho;
 	}
